Change version code of changelog for version 0.19.8
To display changelogs in F-Droid and on their website, the changelog files need to be named like this: `version_code.txt`
As the version codes in the legacy repository are different to these from the main NewPipe repository, rename the latest changelog file for version 0.19.8. This will not have an effect on F-Droid yet, becuase the commit has not been pushed to the 0.19.8 tag.